Reorganized tests to allow for sequential execution for end-to-end tests
